% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_poll_pred.R
\name{stan_poll_pred}
\alias{stan_poll_pred}
\title{Bayesian State-Space Model for Polling Prediction}
\usage{
stan_poll_pred(
  y,
  sigma,
  initial_prior = 50,
  random_walk_sd = 0.2,
  mu_sigma = 1,
  inflator = sqrt(2),
  ...
)
}
\arguments{
\item{y}{a matrix of point estimates}

\item{sigma}{a matrix of polling standard deviations}

\item{random_walk_sd}{with default of \code{0.2}}

\item{mu_sigma}{the average sigma with value of \code{1}}

\item{inflator}{the factor by which to increase error due to non-statistical
polling error}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian State-Space Model for Polling Prediction
}
